Fix "Profiler is not executable" error (#544)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/544

Previously, the same `Makefile` was used (and reused) for building all different kinds of profilers for different ops and op configurations. This has caused an issue when running parallel unit tests in a local environment, as contention for the same `./tmp/profiler/Makefile` has led to different tests rewriting it before being read by others.

As a result, the tests were building each others' profilers and were left without their own. The latter manifested itself in the following error, as the profiler executable that should have been built wouldn't have been there by the time the compilation would have ended:

```
Profiler ./tmp/profiler/gemm_rcr/gemm_rcr_9e46850d5286ecc7e078b5b7f76afbcac62967b4_3 is not executable
```

In this diff, the built profiler target names are included in the per-profiler `Makefile` name, hence excluding the possibility of different tests rewriting each other profiler `Makefile`s. This resolves the issue and the above error is no longer raised. Importantly, it is acceptable for the tests to rewrite the `Makefile` of the same profiler targets, as the content will also be the same.

Additionally, a few retries (with a delay) are made to check if the profiler binary is executable in the `gemm_universal` front-end. This is to handle the cases where the same binary is being compiled in parallel by more than one unit test, so that by the time one tries to check executability, the other is in process of writing the compiled result.

Reviewed By: kadeng

Differential Revision: D44788627

fbshipit-source-id: 2a68d7483d45e8578692bb2f15c35b7a362e0fd6