Add profiler execution retries (#553)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/553

When multiple unit tests are running in parallel, a few can be building the same profiler binary (e.g., for the same op configuration from both tests). In such cases, it may happen that, by the time one test attempts to execute the build profiler binary, another test is in the middle of writing the compilation result. This triggers an error, which before this diff has caused a failure of the async task running profiler commands and eventual profiler timeout.

The diff adds retries to profiler execution, hence remediating the problem described above.

Reviewed By: alexanderguzhva

Differential Revision: D44815907

fbshipit-source-id: c9082e8bc9c59ad1f629373e156ba4661cc89795