Ensure safe build dir name (#561)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/561

In D44642328 we saw a test failure related to some test generating a test name which contains characters which might be unsafe for file or directory names ( e.g. curly brackets in that case ). In order to guard against such a thing, this
diff introduces a small safety check in compile_model which assserts that the test_name argument contains only whitelisted characters ( english alphanumeric, underscore, dash, dot ) and that it is a nonempty string which is not one of the special directories "." or "..".

Differential Revision: D44867941

fbshipit-source-id: 82b758ba92f14effd32264b31683a8d1592e687c