Accept JaggedIntVar in padded_dense_to_jagged
Summary: It may happen that `total_length` passed to the `padded_dense_to_jagged` op is actually a `JaggedIntVar`. In such cases, the `total_length` is fetched from the `shape[0]` of a tensor that already happens to be jagged. Before this diff, this has caused an exception in the `padded_dense_to_jagged` front-end validation. The diff fixes this by fetching the `total_length` from within the passed `JaggedIntVar`.

Differential Revision: D44997496

fbshipit-source-id: 6e39218445b196cba236cf6a48f637282f31eea2