add reshape for `to` converter (#576)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/576

We suppose to bypass this op but in extreme case like
a = placeholder(); return a.to()

It introduces a node in AIT graph which has is_input=True and is_output=True. The node name is output_xx
fx2ait throws error when doing the input name binding. So we add an extra reshape layer here which brings no computation.

Reviewed By: hl475, chenyang78

Differential Revision: D44991256

fbshipit-source-id: afc951c23d205351166a0407ff4b9f218b075bff