Fix race condition in FBCUDA (#591)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/591

A race condition in FBCUDA could be triggered when it was launched in a parallel.
See https://www.internalfb.com/phabricator/paste/view/P697768916

Update:

After further investigation, the source of the race condition appears to be between multiple processes and target_def.py CUDA.__exit__(...) code deleting things, and FBCUDA.__init__(...) code copying and writing files into directories that the other process is busy deleting. It was not obvious that CUDA.__exit__ would delete paths that FBCUDA.__init__() created.

After short discussions, we concluded that there is no way to clean this up with minimal risk code changes - a larger rewrite of target_def.py would be neccessary, also to prevent resource leaks, but is out of scope for a hotfix.

So the safest way to go, and what is now implemented here is the following:

 * Revert target_def.py to the pre-build-cache state where these issues were avoided by creating new randomly named temp directories every time.
 * In order to not get new  build cache keys everytime which would make caching impossible, the build cache makefile normalization was adapted instead.

The impact on build times appears neglible ( approx. 0.08 seconds )

Unit tests were added and modified accordingly as well.

Reviewed By: wushirong, aakhundov

Differential Revision: D45093471

fbshipit-source-id: 9473d96ac33178dcddc98cf988fa4560c7e85f96