Skip incompatible tests outside CI (#606)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/606

Currently, when running the unit tests outside the CI (e.g., on a devgpu or locally), test filtering is not applied. This doesn't create visible issues on A100, as all the current tests are runnable on A100 (or skipped at the class level when they are ROCm-specific). But it does create an issue on e.g. V100 (CC 70) or T4 (CC 75) hosts outside CI: some tests meant to be run on A100 are failing. E.g., see P701981848 for a `test_gemm` run on T4. As we're planning to add H100 / SM90 unit tests in the near future, this will become an issue on A100, too.

This diff extends the test filtering functions in `test_utils.py` to filter out the tests *not runnable* in the local env, even when outside the CI. E.g., `_sm80` and `_bf16` tests are filtered out on V100 and T4, and `_rocm` tests are filtered out everywhere except on ROCm. When `SM90` arch will be added to the `test_utils.py`, only the `_rocm` tests will be skipped in it, and `_sm90` tests will be skipped everywhere else.

Differential Revision: D45217518

fbshipit-source-id: 0b4245092c8da8b8bc98468bddadae21e8d78d57