Add env_variables context manager to test_utils (#607)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/607

A new context manager, `env_variables`, is added to the `test_utils` to facilitate temporary setting (or unsetting) of one or more environment variables for the block of code within the context manager. On exit from the context manager, the state (as set or not set) of the environment variables before entering the context manager is restored.

To temporarily set the environment variable `FIRST_ENV_VAR=1` for the code in `do_something()`:

```
with env_variables(FIRST_ENV_VAR="1"):
    do_something()
```

To temporarily unset (== remove) the environment variable `SECOND_ENV_VAR` for the code in `do_something()`:

```
with env_variables(SECOND_ENV_VAR=None):
    do_something()
```

To do both of the above simultaneously:

```
with env_variables(
    FIRST_ENV_VAR="1",
    SECOND_ENV_VAR=None,
):
    do_something()
```

On exit from the `with env_variables(...): ...` block, both `FIRST_ENV_VAR` and `SECOND_ENV_VAR` are returned to their previous state.

Differential Revision: D45218884

fbshipit-source-id: dcb39637d0c939b46a17301384db2449a92f9068