Set skip_on_empty=True from filter_test_cases_by_params (#609)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/609

Currently, if the test env that the tests are running on (in CI) is not found among the keys of the `params` dict argument of the `filter_test_cases_by_params` function, an empty list will be returned by the function:

https://www.internalfb.com/code/fbsource/[db44bc42edbac0af6b277aca542026005759f338]/fbcode/aitemplate/AITemplate/python/aitemplate/testing/test_utils.py?lines=81

When empty list is passed to the `parameterized.expand` decorator, without additionally passing the `skip_on_empty=True` argument, `parameterized.expand` will raise an exception. This is currently not visible, as the tests are running in the test envs that exist in the `params`. However, if we add ROCm or SM90 CI, this will become a problem.

In this diff, the `filter_test_cases_by_params`function is refactored to return a dict of `kwargs` for the `parameterized.expand` decorator instead of mere list for the first (`input`) argument. The `skip_on_empty=True` argument is included into the returned `kwargs` dict, along with the `input`. This will lead to empty `input` list to be processed normally by `parameterized.expand`.

All call sites of `filter_test_cases_by_params` in the unit tests are prepended with `**` to properly pass the `kwargs` to the decorator. Except a few cases where `itertools.product` is applied on the `input` list: in such cases (e.g., in `test_attention.py`) the value of the `"input"` key is fetched from the dict + `skip_on_empty=True` is added manually.

Reviewed By: chenyang78

Differential Revision: D45219129

fbshipit-source-id: fb12975643b5db261b3be9b6628f41a7cad7cd8f