Refactor test_bmm(_add) with filter_test_cases_by_test_env (#635)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/635

Almost all unit tests from `test_bmm` and `test_bmm_add` are running on *both* V100 and A100 hosts. This is inefficient, as only a small fraction of the tests must run on A100. In this diff, the tests are refactored to rely on `filter_test_cases_by_test_env` instead of `filter_test_cases_by_params`, which leads to a more frugal use of A100 hosts. See the test plan for the before / after numbers.

Reviewed By: hl475

Differential Revision: D45405007

fbshipit-source-id: 98ce1c39d09b057586f759f22b9eedd6bcd4dbd5