Make GEMM ProfilerMemoryPool size computation generic (#653)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/653

Currently there are some hard-coded magic numbers like `(1 << 25)` in computing the `ProfilerMemoryPool`'s size (in terms of the number of fitting largest input sizes). In this diff, the magic number is replaced by the size of the actual L2 cache retrieved from the `device_properties.l2CacheSize`.

Also, the upper bound of the `mem_pool_sz` is increased from 64 to 512. The latter should be safe, given that the upper bound of the allocated total tensor memory per input / output is the size of the L2 cache.

Reviewed By: ipiszy, chenyang78

Differential Revision: D45530460

fbshipit-source-id: ccdf9ded7393ec6adabe61fc6e73d9c98c68bd74