Allow setting jagged tensor total_length upper bound from offsets (#634)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/634

During shape inference in fx2ait, as an alternative to setting the `total_length: IntVar`'s upper bound based on a single globally configured single max sequence length value, it may be set from the content of the correspondibng offsets tensor in the sample input. The latter is more memory efficient in cases when there are multiple jagged tensors with (substantially) different max. sequence lenghts. The assumption is that the offsets tensors in the sample inputs actually contain at least one sequence with the maximum sequence length for the corresponding jagged tensor.

Reviewed By: qxy11, wushirong, tissue3

Differential Revision: D45393984

fbshipit-source-id: 7a35955c2fa28b67ebcfb91925409492d8ad9cd7