Fix get_positive_dim for IntVar (#680)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/680

This is an issue encountered during ig model lowering: https://fb.workplace.com/groups/757073672259175/permalink/903432870956587/
`get_positive_dim` is the function to deduce positive dim for slice_op.
e.g. `a[-2]` where `len(a)=3` denotes we should access `a[1]`
But this only meaningful when the size is IntImm, not dynamic dim.
Now we add the filter so that we do nothing when the size is dynamic dim. Notice this will result in error when idx is negative and result in lowering failure. (This is also true without this filter, basically whenever there is dynamic dim, lowering get error)
This is ideal error, because in that case AIT has undefined behavior and we would want to check the model owner if such lowering happens.

Reviewed By: amateurcoffee

Differential Revision: D45722873

fbshipit-source-id: 0478dceb5b8c69b7801759577a18ea9be851b417