Fix the profiler bug in bmm_xxx_add SM90 kernels (#682)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/682

As `bmm_xxx_add` ops are forming the problem arguments for profiling on their own (in contrast to the `bmm_xxx` ops relying on `bmm_common.default_gen_profiler`), I've missed generation and passing of the CUTLASS 3.x problem arguments for their profiler generation. This has resulted in a really clandestine bug in the profiler that hasn't manifested itself in the CI (because profilers are compiled, which worked, but aren't run, which didn't).

Reviewed By: ipiszy, wushirong

Differential Revision: D45751431

fbshipit-source-id: e30c1e6b327a30812b719835f5b5cfb4827ce1be