Fix a tricky fused_elementwise alignment issue. (#693)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/693

fused_elementwise's max_alignment needs to be adjusted based on input broadcast
properly. Consider the following case:
A: [1, 2, 1]
B: [2, 1, 2]
To calculate A + B, max_alignment needs to be set to 2 instead of 4.
The current implementation doesn't consider the case when broadcast dim is in the
middle. This diff tries to fix it.

This diff also contains a minor refactoring to compute input alignments.

Differential Revision: D45839276

fbshipit-source-id: 4218aada841962714087bae63a860ade12679446