fx2ait: explicitly ensure workdir's existence
Summary:
Right now the only thing that prevent a non-existent workdir from breaking the ait system, is an indirect mkdir call to create the `test_dir`, which is `workdir/ait_name`.

We shouldn't have just replied on that. We should have explicitly make sure workdir exist in the first place.

Reviewed By: amateurcoffee

Differential Revision: D46036294

fbshipit-source-id: 24a9377c5782719487cd692aec578fdc0b310f5c