fx2ait: explicitly ensure workdir's existence (#714)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/714

Right now the only thing that prevent a non-existent workdir from breaking the ait system, is an indirect mkdir call to create the `test_dir`, which is `workdir/ait_name`.

We shouldn't have just replied on that. We should have explicitly make sure workdir exist in the first place.

Reviewed By: amateurcoffee

Differential Revision: D46036294

fbshipit-source-id: 6317a646b1c8c4c1271af8b6a24e2d5e6125259f