Pass bias vector via epilogue schedule in SM90 gemm_rcr_bias (#734)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/734

Since [7dbf423](https://github.com/NVIDIA/cutlass/commit/7dbf42376330230b9c5f0fe2a0ac1c167d1f1889) in `nvidia/cutlass` one can disable the residual and pass the bias vector through the `TmaWarpSpecializedBiasElementwise` epilogue schedule. When the bias vector is column-major, this leads to better performance. For that to work with the `gemm_rcr_bias` op (with the row-major bias and output) we use the transposed problem with swapped A / B and column-major bias and C.

Differential Revision: https://internalfb.com/D46265703

fbshipit-source-id: c881e5010d1d349f9a87a8e4cf47701fc9a47f18