Grouped Classic B2B BMM op 2 ( padded -> jagged operator implementation ) (#737)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/737

Implementation of a jagged (grouped) version of the classic b2b bmm operator.

This diff builds on another diff in the stack, which copied the implementation of the ungrouped classic b2b bmm op so that a new classic b2b bmm operator is created. This diff is about modifying the new operator to implement a jagged / grouped version.

This diff supersedes the diff D45880465, which contains a configurable version including code that turned out not to improve the performance. This code has been removed in this version.

Reviewed By: aakhundov

Differential Revision: D46277337

fbshipit-source-id: 61a8e9d1bce9b23fbcb4f77ce3cd073d4abb3050