Include split+cat in fuse_split optimization (#740)
Summary:
This change extends _fuse_split_and_strided_op to also optimize split followed by cat (when both are on the same dim). The split op is removed and the input_accessors of the cat op are updated.

The new `test_fuse_split_cat.py` test case tests for split+cat across dim 0, dim 1, and a third case where split and cat along different dims (a case that we are not optimizing yet).

Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/740

Reviewed By: wushirong

Differential Revision: D46505169

Pulled By: chenyang78

fbshipit-source-id: 3aa8f58381c2e20ba0967254222b67dc9a106cf1