Standalone test cases (#773)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/773

These changes improve the standalone executable generation of AIT to make the generated standalone test executable a full-fledged unit test and benchmarking tool, supporting loading and testing with test files which specify inputs and expected outputs for models.

This allows for much faster iteration cycles in the development of native operators, especially if combined with the features introduced in the base diff D46519923, which allow in-place development, code navigation and debugging within an AIT build dir.

Reviewed By: aakhundov

Differential Revision: D46649096

fbshipit-source-id: 2f08809073e95b08e86899b572917237136cd281