Package libait_model.so with fx2ait (#815)
Summary:
This PR rewrote the building and loading of `libait_model.so` so that it can be packaged with fx2ait and loaded only once. The code is mostly adopted from `torchvision._C`. After this PR, fx2ait no longer requires the user to put a "./build/libait_model.so" file in the current directory.

One limitation is that it does not support Windows and ROCM platforms.

Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/815

Reviewed By: wushirong

Differential Revision: D47280668

Pulled By: aakhundov

fbshipit-source-id: 4db7b5bcbcab846346b2d2f97e4f1ca8a62283fb