Refactor SimpleDisjointSet in fuse_elementwise to use sets
Summary:
The `SimpleDisjointSet` used in the `fuse_elementwise` transformation pass is using lists to collect groups of the disjoint nodes. In certain cases, this can lead to a heavy duplication in the node groups resulting in the `fuse_elementwise` pass taking much longer time than it should.

This diff refactors the `SimpleDisjointSet` to use sets instead of lists to alleviate the node duplication and the resulting long pass execution.

Reviewed By: sgrigory

Differential Revision: D47466832

fbshipit-source-id: e0550702cfc492b096dbef2b17109aae5d3d7576