Refactor SimpleDisjointSet in fuse_elementwise to use sets (#830)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/830

The `SimpleDisjointSet` used in the `fuse_elementwise` transformation pass is using lists to collect groups of the disjoint nodes. In certain cases, this can lead to a heavy duplication in the node groups resulting in the `fuse_elementwise` pass taking much longer time than it should.

This diff refactors the `SimpleDisjointSet` to use sets instead of lists to alleviate the node duplication and the resulting long pass execution.

Reviewed By: sgrigory

Differential Revision: D47466832

fbshipit-source-id: 13044f18f72763a4de02f769782300b690d7faf3