Have softmax handle any dim via permutation (#832)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/832

Since the backend doesn't yet support reduction over non-last-dimensions, I'm
having the frontend add two permute ops to move the desired dimension to the
end of the shape.

I considered modifying the softmax template code instead of creating new op
instances, but the `permute()` op has a bunch of logic for dispatching to more
specialized versions of itself, and it seems the easiest way to take advantage
of that is by creating instances of the permute op.

Reviewed By: aakhundov

Differential Revision: D47506000

fbshipit-source-id: b3522fdbb0154d9caa5a4865653aec1816de6bc9