Rename reshape to avoid acc tracer collision
Summary:
There is an AIT bug
See this acc_graph: P795462043, and model-generated.h: P795461857
When the node name is the same as function name - in this case, reshape_14
we will have
```
expression preceding parentheses of apparent call must have (pointer-to-) function type
```
This is because reshape_14 is both tensor and function (see model-generated.h: P795461857)
The fix can be either add func name to MEMO set (D47657410), or name our reshape function so that it would cause any name collision to the name produced by the acc tracer (this diff).

We probably chose the latter because the former may might end up with a very large name set and long compilation time.
The later approach isn't general but I guess we don’t have many cases where we name our kernel launchers the same way as what the acc tracer names tensors.

Differential Revision: D47675840

fbshipit-source-id: 21395181fb7610cc5feaccf953dbee0dfc900f51