Support dim != 1 for softmax w/o using permute
Summary:
This is a port of PyTorch's softmax implementation.

Notable differences:
* We use fast_exp & fast_max instead of std::max and std::exp
* We don't use higher-precision types for accumulator values (doesn't look like the dim=-1 softmax code does this either)

This is probably the reason why we are (very marginally) faster.

I have named this new softmax implementation "softmaxGeneral" since it is able to handle arbitrary reduction dimensions, even though we are only using it for the `dim > 1` case.

Differential Revision: D47732875

fbshipit-source-id: 9e9de4a476662958279635be07d90592af55a919