Refactor softmax code for dim=-1 case
Summary:
This sets the stage for extending the warp reduction code to the dim=-2 case.

Main changes in this diff:

1. blockReduceMax now uses `fast_max` instead of `max`. This is what the dim=-2
reduction code already uses. From looking at the implementation, it seems that
fast_max is fast because of type specialization and not because it sacrifices
accuracy, so I think this is a safe change.

2. The block reduction code had a `NUM` parameter that was always set to 1.
I've eliminated that to simplify things.

3. The shared memory used by the block / warp reduction code had shared memory
set to rows of 33 elements in order to avoid bank conflicts. However, given
that NUM was always 1, I don't think bank conflicts are possible in practice.
I've therefore changed the shared memory to use rows of 32 elements instead.

Differential Revision: D47862053

fbshipit-source-id: f4d4e1944cf224e653e8a1dc5d592be6bd12a43f