Support dim != 1 for softmax w/o using permute (#845)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/845

This is a port of PyTorch's softmax implementation.

Notable differences:
* We use fast_exp & fast_max instead of std::max and std::exp
* We don't use higher-precision types for accumulator values (doesn't look like the dim=-1 softmax code does this either)
* We propagate the reduction dim size & inner size as constants

We seem to be very marginally slower than PT for small batch sizes and very marginally faster for large ones.

I have named this new softmax implementation "softmaxGeneral" since it is able to handle arbitrary reduction dimensions, even though we are only using it for the `dim > 1` case.

Reviewed By: aakhundov

Differential Revision: D47732875

fbshipit-source-id: f57fb81921a8f6da48abe5b61542e09f3142e4ed