Better profile_callable
Summary:
I extended profile_callable to accept a list of events of interest. In
the case where this parameter is not specified, profile_callable will collect
the timings of all events except those ending in " Sync". This gives us more
accurate results; we no longer measure overhead from the syncs.

Finally, I converted the softmax benchmark to use profile_callable for both the
AIT and PT2 benchmarking. Using the same profiling function ensures we are
comparing apples to apples.

Differential Revision: D48252668

fbshipit-source-id: cebc4b398e525b2db755192f3c57ab298ac42faf