Avoid input shape mutation in the tile converter (#901)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/901

Currently, the `acc_ops_tile` converter modifies the shape of the input (by `insert`ing values into the mutable list shape). This can causes problems for the downstream fusion passes: e.g., layernorm + sigmoid + mul fusion that relies on the fact that the mul's output is the same rank as the input.

This diff fixes the mutable input shape change in the tile converter by adding an extra reshape op instead.

Differential Revision: D48389978

fbshipit-source-id: c1e370862c0411562c3eb3442d139822f40bad0d