Add fx graph-based jagged tensor identification (#893)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/893

Previously, the input specs (in particular, the `IntVar` dimensions for the total length) of the jagged tensor inputs were inferred based on the value of the total length dimension in the sample jagged tensor inputs. This proved to be brittle, as it's not always easy (or even possible) to guarantee distinct values of the total length for distinct sample jagged tensor inputs.

This diff introduces an alternative way of jagged tensor input spec identification: based on the analysis of the fx graph. More specifically, in the `tag_jagged_tensors` pass, the offsets corresponding to each jagged tensor in the model are identified via the jagged tensor operators (in which jagged tensors and offsets are both inputs) and used to tag the upstream occurrences of the jagged tensors with the same total length in the graph. Further, in the `fx2ait` (`lower.py` / `tensor_spec.py`) the offsets name tags are used to distinguish between different jagged tensor inputs in the model. This also allows identifying different jagged tensors with the same total length (as, for example, in the jagged SHA / MHA example).

Reviewed By: qxy11

Differential Revision: D48227186

fbshipit-source-id: 029841c1e864be1272ced4a3455be5c9bc272f49