Add deduce_jagged_tensor_with_graph_analysis flag for batch dim distinguish (#928)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/928

For vdd, it seems that the jagged tensor batch dim is identical to dense tensor batch dim, which caused issue in bmm kernel, that it cannot handle batch size as large as 2^16.

This fix adds a flag `deduce_jagged_tensor_with_graph_analysis` so that when it is turnt on, we depend on graph analysis, i.e. `try_getting_jagged_tensor_map`, to deduce batch dim for jagged tensor. This can be more reliable than deducing based on value.

Reviewed By: aakhundov

Differential Revision: D49255733

fbshipit-source-id: 0df5df89677231db67c873de653720e63e3fb774