Return early from permute ops on zero-size input
Summary:
Previously with the `permute` ops, we could have run into an edge case when the input to the permute op is a model input. Apparently, zero-size PT tensors return `nullptr` as `.data_ptr()`. This has led to an error in the permute ops' back-end which validates the input pointers to be non-NULL.

In this diff, an early exit condition is added to all permute ops' back-end before the input pointer validation. This remediates the error from the edge case described above.

Differential Revision: https://internalfb.com/D50118462

fbshipit-source-id: 938fc4b4e6138c2237ca16800c8cc4f23f1533da