Return early from permute ops on zero-size input (#946)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/946

Previously with the `permute` ops, we could have run into an edge case when the input to the permute op is a model input. Apparently, zero-size PT tensors return `nullptr` as `.data_ptr()`. This has led to an error in the permute ops' back-end which validates the input pointers to be non-NULL.

In this diff, an early exit condition is added to all permute ops' back-end before the input pointer validation. This remediates the error from the edge case described above.

Reviewed By: qxy11

Differential Revision: D50118462

fbshipit-source-id: 444b55bcb20a875c5aa68fb157a8e5f9a085db6e