Skip fuse_mm_elementwise fusion with model output in the middle (#1000)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/1000

`fuse_mm_elementwise` transformation fuses linear patterns into single ops. As a result, all intermediate outputs in the pattern are eliminated. In a special case when one or more of those intermediate outputs are model outputs, this leads to those model outputs vanishing after the fusion. Here we add skipping the fusion when one of the intermediate outputs in the detected pattern is a model output.

Reviewed By: frank-wei, chenyang78, mengyingdu

Differential Revision: D56340320

fbshipit-source-id: bb0331cc4741929288c2d6ad0d860ab42f149a2b